+++ include/rstl/rmemory_allocator.hpp
@@ -8,7 +8,7 @@
 namespace rstl {
 struct rmemory_allocator {
   template < typename T >
-  static void allocate(T*& out, size_t sz) {
+  static void allocate(T*& out, s32 sz) {
     out = reinterpret_cast< T* >(new u8[sz]);
   }
   template < typename T >
